bot/requirements.txt
# -------------------------
# Torch (CPU, pip-compatible)
# -------------------------
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2


# -------------------------
# Core Web
# -------------------------
fastapi>=0.110,<0.120
starlette>=0.36,<0.40
uvicorn>=0.27,<0.31
google-genai>=1.40

# -------------------------
# Validation / Types
# -------------------------
pydantic>=2.6,<3.0
typing-extensions>=4.9

# -------------------------
# Async / HTTP
# -------------------------
httpx>=0.26,<0.29
anyio>=4.3,<5.0
sniffio>=1.3

# -------------------------
# Serialization / Speed
# -------------------------
orjson>=3.9
ujson>=5.9

# -------------------------
# ML Core (CPU-only, wheels available)
# -------------------------
numpy>=1.26,<2.3
scipy>=1.11,<1.16
scikit-learn>=1.3,<1.8


# -------------------------
# Transformers / RAG
# -------------------------
transformers>=4.38,<4.58
tokenizers>=0.15,<0.23
huggingface-hub>=0.20,<1.0
sentence-transformers>=2.6,<6.0
chromadb>=0.4.24,<1.5

# -------------------------
# Utilities
# -------------------------
requests>=2.31
tqdm>=4.66
regex>=2023.12
python-dotenv>=1.0
psutil>=5.9

# -------------------------
# Security / Auth
# -------------------------
bcrypt>=4.1
google-auth>=2.28
oauthlib>=3.2
requests-oauthlib>=1.3

# -------------------------
# Parsing / Files
# -------------------------
PyYAML>=6.0
python-dateutil>=2.8
lxml>=4.9
pillow>=10.2
#-------------------
firebase
#--------------------------
firebase-admin>=6.4.0
google-cloud-firestore>=2.14.0
